First pass at parameter passing / log generation.
The idea is that every task's parameters give a unique id (via SHA3)
that can be used to reliably generate unique log names. These logs,
which contain data as rows of JSON objects, can then be read in and
processed to derive the parameters of future tasks.

Two kinds of parameter passing are given (though only one is tested at
the moment):

 - literal, which does direct substitution on the command line
 - path, which writes the contents to an unnamed temporary file and
 passes the FD (as /proc/<pid>/fd/<fd>) to the task. Note that this only
 works on linux. At this time, I've got no reason to change that.

Note that even when path passing is used, the parameter *value* is still
used to compute the task identifier (ie its parameter hash).

Note also that the path of the program being run in a task is used as
input. Using absolute paths will, obviously, prevent tasks from being
identified correctly across machines. I am not sure what to do about
this.